project_IRMP_code/requirements.txt
Click==7.0
Flask~=0.12.5
Flask-FlatPages==0.7.1
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
Pygments==2.4.2
PyYAML==5.1.2
Werkzeug==0.16.0
flask-login
Flask-WTF
utm~=0.7.0
geopy~=2.3.0
Pillow~=9.2.0
folium~=0.14.0
branca~=0.6.0
SQLAlchemy~=2.0.11
numpy~=1.19.3